% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms.R
\name{as_storms}
\alias{as_storms}
\title{Convert POINT collection to LINESTRING by storm}
\usage{
as_storms(x = read_hurdat())
}
\arguments{
\item{x}{hurdat POINT data for one or more storms}
}
\value{
LINESTRING object, one per storm
}
\description{
Convert POINT collection to LINESTRING by storm
}
