% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_hurdat}
\alias{fetch_hurdat}
\title{Fetch HURDAT2 data}
\usage{
fetch_hurdat(url = hurdat2_url(), dest = hurdat_path("raw"))
}
\arguments{
\item{url}{chr, the URL of the source data}

\item{dest}{chr, the destination path}
}
\value{
the output of \code{\link[utils:download.file]{utils::download.file()}}, 0 for success
}
\description{
Fetch HURDAT2 data
}
