% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat2-package.R
\docType{package}
\name{hurdat2}
\alias{hurdat2-package}
\alias{hurdat2}
\alias{hurdat}
\title{hurdat2}
\description{
Serves a subset of ecopmo jellyfish predictive maps
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/hurdat2}
}

}
\keyword{internal}
