% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{hurdat_path}
\alias{hurdat_path}
\title{Retrieve the root data path, possibly adding segments}
\usage{
hurdat_path(..., path = root_data_path())
}
\arguments{
\item{...}{chr, path segments and/or file names to append to the \code{path}}

\item{path}{chr, the root data path}
}
\value{
path description
}
\description{
Retrieve the root data path, possibly adding segments
}
