% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_hurdat}
\alias{list_hurdat}
\title{List HURDAT2 data file(s)}
\usage{
list_hurdat(path = hurdat_path("raw"))
}
\arguments{
\item{path}{chr the HURDAT2 data path}
}
\value{
chr, vector of one or more HURDAT2 data file descriptions
}
\description{
List HURDAT2 data file(s)
}
