% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_coast}
\alias{read_coast}
\title{Read coastline from Natural Earth}
\usage{
read_coast(
  filename = system.file("extdata/coastline.rds", package = "hurdat2")
)
}
\arguments{
\item{filename}{chr, the name of the file}
}
\value{
sfc_MULTILINESTRING geometry
}
\description{
Made with Natural Earth. Free vector and raster map data @ naturalearthdata.com.
}
