% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{record_identifier}
\alias{record_identifier}
\alias{system_status}
\title{Retrieve a look-up table for record_id and sys_status}
\usage{
record_identifier()

system_status()
}
\value{
named character vector
}
\description{
Retrieve a look-up table for record_id and sys_status
}
