% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_root_data_path}
\alias{set_root_data_path}
\alias{root_data_path}
\title{Set or set the root data path}
\usage{
set_root_data_path(
  path = "/mnt/s1/projects/ecocast/coredata/noaa/nhc/hurdat2",
  filename = "~/.hurdat2"
)

root_data_path(filename = "~/.hurdat2")
}
\arguments{
\item{path}{the data path}

\item{filename}{chr, the name of the file where the path is stored}
}
\value{
the input path
}
\description{
Set or set the root data path
}
