% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hurdat_raw}
\alias{read_hurdat_raw}
\alias{read_hurdat}
\title{Read HURDAT2 data}
\usage{
read_hurdat_raw(filename = tail(list_hurdat(), n = 1))

read_hurdat(form = c("point", "linestring")[1])
}
\arguments{
\item{filename}{the name of the file}

\item{form}{chr, one of "point" (default) or "linestring".  The latter
collapses point data into one-row-per-storm linestrings}
}
\value{
\code{read_hurdat_raw} returns a data frame while \code{read_hurdat} returns
an sf POINT data frame
}
\description{
Read HURDAT2 data
}
\seealso{
\href{https://www.aoml.noaa.gov/hrd/hurdat/hurdat2-format.pdf}{AOML format description}
}
