% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hurdat_raw}
\alias{read_hurdat_raw}
\title{Read HURDAT2 data}
\usage{
read_hurdat_raw(filename = tail(list_hurdat(), n = 1))
}
\arguments{
\item{filename}{the name of the file}
}
\value{
data frame
}
\description{
Read HURDAT2 data
}
\seealso{
\href{https://www.aoml.noaa.gov/hrd/hurdat/hurdat2-format.pdf}{AOML format description}
}
